pyproject.toml
[tool.poetry]
name = "Social Media Tracker"
version = "0.1.0"
description = "Social Media Sentiment Tracker for Global Trends"
authors = ["F. Bossart <florent.bossart@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.09,<3.13"
apache-airflow = "^2.10.4"
psycopg2-binary = "^2.9.10"
apache-airflow-providers-postgres = "^5.14.0"
dbt-core = "^1.9.1"
pendulum = "^3.0.0"
dbt-postgres = "^1.9.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.1"
python-dotenv = "^1.1.0"
praw = "^7.8.1"
requests = "^2.32.3"
google-api-python-client = "^2.169.0"
trove-classifiers = "^2025.4.28.22"
transformers = "^4.52.1"
sentencepiece = "^0.2.0"
torch = "^2.7.0"
sentence-transformers = "^4.1.0"
spacy = "^3.8.6"
ginza = "^5.2.0"
ja-ginza = "^5.2.0"
sqlalchemy = ">=1.4.36,<2.0"
pandas = "1.5.3"
numpy = "<1.27"
apache-airflow-providers-docker = "^4.4.0"
sacremoses = "^0.1.1"
streamlit = {version = "^1.45.1", python = ">=3.9,<3.9.7 || >3.9.7,<3.13"}
wordcloud = "^1.9.4"
psutil = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
apache-airflow-providers-sqlite = "^4.0.0"
dockerfile = "^3.3.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = ["optional: not required, nor saved in test_output.txt"]
